workershub/backend/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
django-filter==25.1
mysqlclient==2.2.7
Pillow==11.3.0
python-decouple==3.8
asgiref==3.9.1
sqlparse==0.5.3